feat(mod): add output module and skeleton types
